Fix deletion of posts from the database

+++ api/controllers/post.controller.js
@@ -88,9 +88,17 @@ export const deletePost = async (req, res) => {
       where: { id },
     });
 
+    if (!post) {
+      return res.status(404).json({ message: "Post not found!" });
+    }
+
     if (post.userId !== tokenUserId) {
       return res.status(403).json({ message: "Not Authorized!" });
     }
+
+    await prisma.postDetail.deleteMany({
+      where: { postId: id },
+    });
 
     await prisma.post.delete({
       where: { id },
